requirements.txt
requests == 2.32.3
mysql-connector-python == 9.0.0 
python-dotenv == 1.0.1
pytesseract == 0.3.13
pandas == 2.2.2
numpy == 2.1.1
matplotlib == 3.9.2
ipython == 8.27.0
playwright == 1.46.0
pillow == 10.4.0
beautifulsoup4 == 4.12.3